// src/types/student.tsx
export interface Student {
  id: number
  name: string
  hiragana: string
  birth: string
  phone: string
  post_number: string
  address: string
  grade: number
  class: number
  attend_number: number
  house: string
  extra_activity: string
  curriculum: string
  enrolled_sibling: string
  parent: string
  parent_phone: string
}